requirements.txt
#BASE
#====
Django==1.8
#Django==1.11.20
django-ajax-selects==1.4.3
django-allauth==0.31.0
django-crispy-forms==1.6.0
django-mobile==0.6.0
inflection==0.3.1
Pillow==3.1.0
#psycopg2==2.6.1
python-openid==2.2.5
psycopg2==2.7.1
requests==2.13.0
requests-oauthlib==0.8.0
sqlparse==0.2.4
Werkzeug==0.11.11

#DEVELOPMENT
#============
#django-debug-toolbar==1.9.1
#wheel==0.24.0

#PRODUCTION
#===========
boto3==1.9.137
botocore==1.12.137
dj-database-url==0.5.0
django-storages==1.7.1
docutils==0.14
futures==3.2.0
gunicorn==19.9.0
jmespath==0.9.4
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.2
whitenoise==4.1.2